<==== Mercurial data ====>
Branch: master

+++ edge_firmware/rpi/maven/filter-resources/create_mediaserver_tar.sh
@@ -112,8 +112,10 @@ mkdir -p $BUILD_DIR/$PREFIX_DIR/$MODULE_NAME/bin/
 cp $BUILD_OUTPUT_DIR/bin/release/mediaserver $BUILD_DIR/$PREFIX_DIR/$MODULE_NAME/bin/
 
 #copying plugins
-mkdir -p $BUILD_DIR/$PREFIX_DIR/$MODULE_NAME/bin/plugins
-cp $BUILD_OUTPUT_DIR/bin/release/plugins/*.* $BUILD_DIR/$PREFIX_DIR/$MODULE_NAME/bin/plugins/
+if [ -e "$BUILD_OUTPUT_DIR/bin/release/plugins" ]; then
+  mkdir -p $BUILD_DIR/$PREFIX_DIR/$MODULE_NAME/bin/plugins
+  cp $BUILD_OUTPUT_DIR/bin/release/plugins/*.* $BUILD_DIR/$PREFIX_DIR/$MODULE_NAME/bin/plugins/
+fi
 
 #conf
 mkdir -p $BUILD_DIR/$PREFIX_DIR/$MODULE_NAME/etc/
